Changing the interface for dimensional traits and getting to work on compact kd tree data structure.
The previous interface for dimension traits required the user to
do some metaprogramming himself, which made the definitions for user
defined type kd trees quite bad to work with. The changes still require
the user to supply information on the number of dimensions on a traits class
but other than that he just has to define get functions for his types.
TODO: see if there is a way to automatically detect the upmost index
for the user defined get function.

Compact KD Tree is a similar kd tree than the one we have implemented
in the project, with the difference it stores the elements in all nodes
and it is stored as an array of elements instead of pointers. But, in
contrast with our current implementation, which can support insertions
although it degrades the performance of the tree, it is a dense data
structure, meaning that adding elements would be a disaster.

+++ include/data_structure/multidimensional/quad_tree.hpp
@@ -113,7 +113,7 @@ namespace tcc {
         assert( _3 == first_child + 3 && "Bottom right child should be allocated 3 after top left.\n" );
         int linked_list_element = cur_node->m_element_index;
         while( linked_list_element != -1 ) {
-          assert( m_elements[ linked_list_element ].m_node_element < m_data.size() && "Node element should be a valid index\n" );
+          assert( m_elements[ linked_list_element ].m_node_element < ( int ) m_data.size() && "Node element should be a valid index\n" );
           int32_t element_index = m_elements[ linked_list_element ].m_node_element;
           auto& element_rect = m_data[ m_elements[ linked_list_element ].m_node_element ].second;
           __insert_into_quadrants__( first_child, node_rect, element_index, element_rect, depth );
